pyproject.toml
[tool.poetry]
name = "aichat"
version = "0.1.0"
description = ""
authors = ["Xaron11 <kaczmarzykdawid4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SpeechRecognition = "^3.8.1"
pyttsx3 = "^2.90"
PyQt5 = "^5.15.4"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
